requirements.txt
telethon
pyyaml
discord.py
aiohttp